+++ lib/libsesame3bt-core/ServerCoreImpl.cpp
@@ -4,6 +4,7 @@
 #include "Sesame.h"
 #include "debug.h"
 #include "hal.h"
+#include "libsesame3bt/ClientCore.h"
 #include "libsesame3bt/util.h"
 
 namespace libsesame3bt::core {
@@ -240,18 +241,23 @@ SesameServerCoreImpl::handle_cmd_with_tag(ServerSession& session, Sesame::item_c
 	}
 	std::optional<history_tag_type_t> trigger_type;
 	std::string tstr;
+	float scaled_voltage = NAN;
 	if (std::to_integer<uint8_t>(payload[0]) > 0) {
 		tstr = std::string(reinterpret_cast<const char*>(payload + 1), std::to_integer<size_t>(payload[0]));
 	} else if (size >= 18) {
 		trigger_type = static_cast<history_tag_type_t>(payload[1]);
 		tstr = util::bin2hex(payload + 2, 16);
+		if (size >= 20) {
+			uint16_t voltage_raw = static_cast<uint8_t>(payload[19]) << 8 | static_cast<uint8_t>(payload[18]);
+			scaled_voltage = Status::status_value_to_scaled_voltage_os3(voltage_raw);
+		}
 	} else {
 		tstr = {};
 	}
 	DEBUG_PRINTLN("cmd=%s(%s)", cmd_string(cmd), tstr.c_str());
 	Sesame::response_os3_t res;
 	if (on_command_callback) {
-		res.result = on_command_callback(session.session_id, cmd, tstr, trigger_type);
+		res.result = on_command_callback(session.session_id, cmd, tstr, trigger_type, scaled_voltage);
 	} else {
 		res.result = Sesame::result_code_t::not_supported;
 	}