+++ lib/libsesame3bt-core/os3.cpp
@@ -182,6 +182,10 @@ OS3Handler::handle_history(const std::byte* in, size_t in_len) {
 			auto str = util::bin2hex(tag_data + 2, 16);
 			history.tag_len = str.length();
 			std::copy(str.cbegin(), str.cend(), history.tag);
+			if (in_len >= sizeof(Sesame::response_history_5_t) + 20) {
+				uint16_t voltage_raw = static_cast<uint8_t>(tag_data[19]) << 8 | static_cast<uint8_t>(tag_data[18]);
+				history.scaled_voltage = Status::status_value_to_scaled_voltage_os3(voltage_raw);
+			}
 		}
 	}
 	history.type = histtype;